Quality sweep: security, a11y, cohere, responsive, tidy
- Fix SQL injection in CRM getLeads: replace sql.raw() with parameterized queries
- Fix throw Error → TRPCError in team.ts for proper error contract
- Add ::int cast to count(*) in admin.ts and analytics.ts
- Add dialog ARIA (role, aria-modal, aria-labelledby) to TaskModal
- Add htmlFor/id label associations to all TaskModal form fields
- Add aria-label to all icon-only buttons (close, edit, delete, drag, toolbar)
- Add aria-label to filter selects, aria-pressed to view toggles
- Increase task card touch targets to 44px on mobile
- Fix communities tree: always-visible actions on touch, cap indent depth
- Fix gallery.getBySlug to throw NOT_FOUND instead of returning undefined
- Remove duplicate getTeamMembers from CRM router
- Remove dead posts table, postRouter, and SQLite/libSQL fallback
- Remove 3 unused imports (Lock, useCallback, searchParams)
- Add missing SUPABASE_SERVICE_ROLE_KEY and ANTHROPIC_API_KEY to .env.example
- Update STATUS.md and TODO.md for completed Tasks feature

+++ src/server/api/routers/admin.ts
@@ -21,13 +21,13 @@ export const adminRouter = createTRPCRouter({
       recentQuestionnaires,
       recentWaivers,
     ] = await Promise.all([
-      ctx.db.select({ count: sql<number>`count(*)` }).from(contacts),
+      ctx.db.select({ count: sql<number>`count(*)::int` }).from(contacts),
       ctx.db
-        .select({ count: sql<number>`count(*)` })
+        .select({ count: sql<number>`count(*)::int` })
         .from(questionnaireResponses),
-      ctx.db.select({ count: sql<number>`count(*)` }).from(eventWaivers),
-      ctx.db.select({ count: sql<number>`count(*)` }).from(contactActivities),
-      ctx.db.select({ count: sql<number>`count(*)` }).from(contactSources),
+      ctx.db.select({ count: sql<number>`count(*)::int` }).from(eventWaivers),
+      ctx.db.select({ count: sql<number>`count(*)::int` }).from(contactActivities),
+      ctx.db.select({ count: sql<number>`count(*)::int` }).from(contactSources),
       ctx.db.query.contacts.findMany({
         orderBy: [desc(contacts.createdAt)],
         limit: 5,
@@ -72,7 +72,7 @@ export const adminRouter = createTRPCRouter({
           limit: input.limit,
         });
         const total = await ctx.db
-          .select({ count: sql<number>`count(*)` })
+          .select({ count: sql<number>`count(*)::int` })
           .from(contacts);
         return { data, total: total[0]?.count ?? 0 };
       }
@@ -83,7 +83,7 @@ export const adminRouter = createTRPCRouter({
           limit: input.limit,
         });
         const total = await ctx.db
-          .select({ count: sql<number>`count(*)` })
+          .select({ count: sql<number>`count(*)::int` })
           .from(questionnaireResponses);
         return { data, total: total[0]?.count ?? 0 };
       }
@@ -94,7 +94,7 @@ export const adminRouter = createTRPCRouter({
         limit: input.limit,
       });
       const total = await ctx.db
-        .select({ count: sql<number>`count(*)` })
+        .select({ count: sql<number>`count(*)::int` })
         .from(eventWaivers);
       return { data, total: total[0]?.count ?? 0 };
     }),
