+++ src/server/api/routers/analytics.ts
@@ -67,10 +67,10 @@ export const analyticsRouter = createTRPCRouter({
       totalEmailLinks,
       recentContacts,
     ] = await Promise.all([
-      ctx.db.select({ count: sql<number>`count(*)` }).from(contacts),
-      ctx.db.select({ count: sql<number>`count(*)` }).from(sessions),
-      ctx.db.select({ count: sql<number>`count(*)` }).from(events),
-      ctx.db.select({ count: sql<number>`count(*)` }).from(emailLinks),
+      ctx.db.select({ count: sql<number>`count(*)::int` }).from(contacts),
+      ctx.db.select({ count: sql<number>`count(*)::int` }).from(sessions),
+      ctx.db.select({ count: sql<number>`count(*)::int` }).from(events),
+      ctx.db.select({ count: sql<number>`count(*)::int` }).from(emailLinks),
       ctx.db.query.contacts.findMany({
         orderBy: [desc(contacts.createdAt)],
         limit: 10,
